Moved StackTraceAnalyzer to the `analyzers` folder
