+++ src/analyzers/stack-trace-analyzer/callee-data-extractors/FunctionDeclarationCalleeDataExtractor.ts
@@ -3,32 +3,28 @@ import { injectable } from 'inversify';
 import * as estraverse from 'estraverse';
 import * as ESTree from 'estree';
 
-import { ICalleeData } from '../../interfaces/stack-trace-analyzer/ICalleeData';
+import { ICalleeData } from '../../../interfaces/analyzers/stack-trace-analyzer/ICalleeData';
 
 import { AbstractCalleeDataExtractor } from './AbstractCalleeDataExtractor';
-import { Node } from '../../node/Node';
-import { NodeUtils } from '../../node/NodeUtils';
+import { Node } from '../../../node/Node';
+import { NodeUtils } from '../../../node/NodeUtils';
 
 @injectable()
-export class FunctionExpressionCalleeDataExtractor extends AbstractCalleeDataExtractor {
+export class FunctionDeclarationCalleeDataExtractor extends AbstractCalleeDataExtractor {
     /**
      * @param {Node[]} blockScopeBody
      * @param {Identifier} callee
      * @returns {ICalleeData}
      */
     public extract (blockScopeBody: ESTree.Node[], callee: ESTree.Identifier): ICalleeData|null {
-        let calleeBlockStatement: ESTree.BlockStatement|null = null;
-
-        if (Node.isIdentifierNode(callee)) {
-            calleeBlockStatement = this.getCalleeBlockStatement(
-                NodeUtils.getBlockScopesOfNode(blockScopeBody[0])[0],
-                callee.name
-            );
+        if (!Node.isIdentifierNode(callee)) {
+            return null;
         }
 
-        if (Node.isFunctionExpressionNode(callee)) {
-            calleeBlockStatement = callee.body;
-        }
+        const calleeBlockStatement: ESTree.BlockStatement|null = this.getCalleeBlockStatement(
+            NodeUtils.getBlockScopesOfNode(blockScopeBody[0])[0],
+            callee.name
+        );
 
         if (!calleeBlockStatement) {
             return null;
@@ -36,7 +32,7 @@ export class FunctionExpressionCalleeDataExtractor extends AbstractCalleeDataExt
 
         return {
             callee: calleeBlockStatement,
-            name: callee.name || null
+            name: callee.name
         };
     }
 
@@ -49,13 +45,8 @@ export class FunctionExpressionCalleeDataExtractor extends AbstractCalleeDataExt
         let calleeBlockStatement: ESTree.BlockStatement|null = null;
 
         estraverse.traverse(targetNode, {
-            enter: (node: ESTree.Node, parentNode: ESTree.Node): any => {
-                if (
-                    Node.isFunctionExpressionNode(node) &&
-                    Node.isVariableDeclaratorNode(parentNode) &&
-                    Node.isIdentifierNode(parentNode.id) &&
-                    parentNode.id.name === name
-                ) {
+            enter: (node: ESTree.Node): any => {
+                if (Node.isFunctionDeclarationNode(node) && node.id.name === name) {
                     calleeBlockStatement = node.body;
 
                     return estraverse.VisitorOption.Break;
