+++ src/analyzers/stack-trace-analyzer/callee-data-extractors/FunctionExpressionCalleeDataExtractor.ts
@@ -3,28 +3,32 @@ import { injectable } from 'inversify';
 import * as estraverse from 'estraverse';
 import * as ESTree from 'estree';
 
-import { ICalleeData } from '../../interfaces/stack-trace-analyzer/ICalleeData';
+import { ICalleeData } from '../../../interfaces/analyzers/stack-trace-analyzer/ICalleeData';
 
 import { AbstractCalleeDataExtractor } from './AbstractCalleeDataExtractor';
-import { Node } from '../../node/Node';
-import { NodeUtils } from '../../node/NodeUtils';
+import { Node } from '../../../node/Node';
+import { NodeUtils } from '../../../node/NodeUtils';
 
 @injectable()
-export class FunctionDeclarationCalleeDataExtractor extends AbstractCalleeDataExtractor {
+export class FunctionExpressionCalleeDataExtractor extends AbstractCalleeDataExtractor {
     /**
      * @param {Node[]} blockScopeBody
      * @param {Identifier} callee
      * @returns {ICalleeData}
      */
     public extract (blockScopeBody: ESTree.Node[], callee: ESTree.Identifier): ICalleeData|null {
-        if (!Node.isIdentifierNode(callee)) {
-            return null;
+        let calleeBlockStatement: ESTree.BlockStatement|null = null;
+
+        if (Node.isIdentifierNode(callee)) {
+            calleeBlockStatement = this.getCalleeBlockStatement(
+                NodeUtils.getBlockScopesOfNode(blockScopeBody[0])[0],
+                callee.name
+            );
         }
 
-        const calleeBlockStatement: ESTree.BlockStatement|null = this.getCalleeBlockStatement(
-            NodeUtils.getBlockScopesOfNode(blockScopeBody[0])[0],
-            callee.name
-        );
+        if (Node.isFunctionExpressionNode(callee)) {
+            calleeBlockStatement = callee.body;
+        }
 
         if (!calleeBlockStatement) {
             return null;
@@ -32,7 +36,7 @@ export class FunctionDeclarationCalleeDataExtractor extends AbstractCalleeDataEx
 
         return {
             callee: calleeBlockStatement,
-            name: callee.name
+            name: callee.name || null
         };
     }
 
@@ -45,8 +49,13 @@ export class FunctionDeclarationCalleeDataExtractor extends AbstractCalleeDataEx
         let calleeBlockStatement: ESTree.BlockStatement|null = null;
 
         estraverse.traverse(targetNode, {
-            enter: (node: ESTree.Node): any => {
-                if (Node.isFunctionDeclarationNode(node) && node.id.name === name) {
+            enter: (node: ESTree.Node, parentNode: ESTree.Node): any => {
+                if (
+                    Node.isFunctionExpressionNode(node) &&
+                    Node.isVariableDeclaratorNode(parentNode) &&
+                    Node.isIdentifierNode(parentNode.id) &&
+                    parentNode.id.name === name
+                ) {
                     calleeBlockStatement = node.body;
 
                     return estraverse.VisitorOption.Break;
