+++ src/custom-nodes/AbstractCustomNodeGroup.ts
@@ -7,7 +7,7 @@ import { ICustomNode } from '../interfaces/custom-nodes/ICustomNode';
 import { ICustomNodeGroup } from '../interfaces/custom-nodes/ICustomNodeGroup';
 import { IOptions } from '../interfaces/options/IOptions';
 import { IRandomGenerator } from '../interfaces/utils/IRandomGenerator';
-import { IStackTraceData } from '../interfaces/stack-trace-analyzer/IStackTraceData';
+import { IStackTraceData } from '../interfaces/analyzers/stack-trace-analyzer/IStackTraceData';
 
 import { CustomNode } from '../enums/container/custom-nodes/CustomNode';
 import { ObfuscationEvent } from '../enums/event-emitters/ObfuscationEvent';
