+++ src/interfaces/analyzers/IAnalyzer.d.ts
@@ -1,0 +1,9 @@
+import * as ESTree from 'estree';
+
+export interface IAnalyzer {
+    /**
+     * @param blockScopeBody
+     * @returns any
+     */
+    analyze (blockScopeBody: ESTree.Node[]): any;
+}
