+++ src/interfaces/analyzers/stack-trace-analyzer/ICalleeData.d.ts
