+++ src/interfaces/analyzers/stack-trace-analyzer/ICalleeDataExtractor.d.ts
