+++ src/interfaces/analyzers/stack-trace-analyzer/IStackTraceAnalyzer.d.ts
@@ -1,8 +1,9 @@
 import * as ESTree from 'estree';
 
+import { IAnalyzer } from '../IAnalyzer';
 import { IStackTraceData } from './IStackTraceData';
 
-export interface IStackTraceAnalyzer {
+export interface IStackTraceAnalyzer extends IAnalyzer {
     /**
      * @param blockScopeBody
      * @returns IStackTraceData[]
