+++ src/interfaces/analyzers/stack-trace-analyzer/IStackTraceData.d.ts
