+++ src/types/analyzers/stack-trace-analyzer/TObjectMembersCallsChain.d.ts
