+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/basic-1.js
