+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/basic-2.js
