+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/call-before-declaration.js
