+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/call-expression-of-object-member-1.js
