+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/deep-conditions-nesting.js
