+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/no-call-expressions.js
