+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/no-recursion.js
