+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/only-call-expression.js
