+++ test/functional-tests/analyzers/stack-trace-analyzer/fixtures/self-invoking-functions.js
